THETA プラグイン設定画面の Android 10 対応。
