+++ dConnectDevicePlugin/dConnectDeviceTheta/app/src/main/java/org/deviceconnect/android/deviceplugin/theta/ThetaDeviceService.java
@@ -6,15 +6,26 @@
  */
 package org.deviceconnect.android.deviceplugin.theta;
 
+import android.annotation.TargetApi;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.net.ConnectivityManager;
+import android.net.Network;
+import android.net.NetworkCapabilities;
 import android.net.NetworkInfo;
+import android.net.NetworkRequest;
+import android.net.wifi.ScanResult;
 import android.net.wifi.WifiInfo;
 import android.net.wifi.WifiManager;
+import android.net.wifi.WifiNetworkSpecifier;
 import android.os.Build;
+import android.os.Handler;
+import android.os.Looper;
+import android.support.annotation.NonNull;
 import android.util.Log;
+import android.widget.Toast;
 
 import com.theta360.lib.PtpipInitiator;
 import com.theta360.lib.ThetaException;
@@ -22,7 +33,10 @@ import com.theta360.lib.ThetaException;
 import org.deviceconnect.android.deviceplugin.theta.core.ThetaDevice;
 import org.deviceconnect.android.deviceplugin.theta.core.ThetaDeviceClient;
 import org.deviceconnect.android.deviceplugin.theta.core.ThetaDeviceEventListener;
+import org.deviceconnect.android.deviceplugin.theta.core.ThetaDeviceException;
+import org.deviceconnect.android.deviceplugin.theta.core.ThetaDeviceFactory;
 import org.deviceconnect.android.deviceplugin.theta.core.ThetaDeviceManager;
+import org.deviceconnect.android.deviceplugin.theta.core.ThetaV;
 import org.deviceconnect.android.deviceplugin.theta.core.wifi.WifiStateEventListener;
 import org.deviceconnect.android.deviceplugin.theta.profile.ThetaMediaStreamRecordingProfile;
 import org.deviceconnect.android.deviceplugin.theta.profile.ThetaOmnidirectionalImageProfile;
@@ -37,9 +51,13 @@ import org.deviceconnect.android.profile.SystemProfile;
 import org.deviceconnect.android.provider.FileManager;
 import org.deviceconnect.android.service.DConnectService;
 
+import java.io.IOException;
+import java.net.Socket;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
+
+import javax.net.SocketFactory;
 
 /**
  * Theta Device Service.
@@ -48,6 +66,10 @@ import java.util.logging.Logger;
  */
 public class ThetaDeviceService extends DConnectMessageService
     implements ThetaDeviceEventListener {
+
+    public static final String ACTION_CONNECT_WIFI = "action.CONNECT_WIFI";
+
+    public static final String EXTRA_SCAN_RESULT = "scanResult";
 
     /** ロガー. */
     private final Logger mLogger = Logger.getLogger("theta.dplugin");
@@ -79,6 +101,17 @@ public class ThetaDeviceService extends DConnectMessageService
             }
         }
     };
+
+    @Override
+    public int onStartCommand(final Intent intent, final int flags, final int startId) {
+        int result = super.onStartCommand(intent, flags, startId);
+        if (intent != null && ACTION_CONNECT_WIFI.equals(intent.getAction())) {
+            ScanResult scanResult = intent.getParcelableExtra(EXTRA_SCAN_RESULT);
+            connectWifi(scanResult);
+        }
+        return result;
+    }
+
     @Override
     public void onCreate() {
         super.onCreate();
@@ -201,4 +234,43 @@ public class ThetaDeviceService extends DConnectMessageService
         }
     }
 
+    @TargetApi(Build.VERSION_CODES.Q)
+    private void connectWifi(final ScanResult result) {
+        WifiNetworkSpecifier specifier = new WifiNetworkSpecifier.Builder()
+                .setSsid(result.SSID)
+                .setWpa2Passphrase(parsePassword(result.SSID))
+                .build();
+        NetworkRequest request = new NetworkRequest.Builder()
+                .addTransportType(NetworkCapabilities.TRANSPORT_WIFI)
+                .removeCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+                .setNetworkSpecifier(specifier)
+                .build();
+        requestNetwork(request);
+    }
+
+    @TargetApi(Build.VERSION_CODES.Q)
+    private boolean requestNetwork(final NetworkRequest request) {
+        ConnectivityManager connectivityManager = getConnectivityManager();
+        if (connectivityManager == null) {
+            return false;
+        }
+        final ConnectivityManager.NetworkCallback callback = new ConnectivityManager.NetworkCallback() {
+            @Override
+            public void onAvailable(final @NonNull Network network) {
+                super.onAvailable(network);
+
+                Toast.makeText(getApplicationContext(), "onAvailable", Toast.LENGTH_LONG).show();
+            }
+        };
+        connectivityManager.requestNetwork(request, callback, new Handler(Looper.getMainLooper()));
+        return true;
+    }
+
+    private String parsePassword(final String ssid) {
+        return ssid.substring(7, 7 + 8);
+    }
+
+    private ConnectivityManager getConnectivityManager() {
+        return (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+    }
 }
