+++ dConnectDevicePlugin/dConnectDeviceTheta/app/src/main/java/org/deviceconnect/android/deviceplugin/theta/activity/ThetaDeviceActivity.java
@@ -77,6 +77,7 @@ public class ThetaDeviceActivity extends FragmentActivity {
         filter.addAction(WifiManager.WIFI_STATE_CHANGED_ACTION);
         registerReceiver(mWifiReceiver, filter);
     }
+
     @Override
     public void onDestroy() {
         super.onDestroy();
